Improve E2E testing
Move all "test:package" logic into "test:e2e" and make "test:e2e" used to test the built packages
Removes previous "test:e2e" that used container built during testing, instead rely on beaker managed container that uses RPMs built during testing to test e2e functionality

Fixes #1351
Fixes #1341
