--- Dockerfile.test
@@ -1,17 +1,0 @@
-FROM ood:latest
-
-ARG USER=ood
-ARG UID=1000
-ARG GID=1000
-
-RUN dnf install -y \
-        python3 && \
-    dnf clean all && rm -rf /var/cache/dnf/* && \
-    python3 -m pip \
-        install flask
-
-RUN groupadd -g $GID $USER && \
-    useradd -u $UID --create-home --gid $USER $USER
-
-CMD [ "/opt/ood/launch" ]
-
