+++ spec/e2e/pun_pre_hook_spec.rb
@@ -1,49 +1,64 @@
-require 'watir'
-require_relative 'e2e_helper'
+require 'spec_helper_e2e'
 
 describe 'Pun Pre Hook' do
-
   def browser
     @browser ||= new_browser
   end
 
-  after(:all) do
-    Rake::Task['test:stop_test_container'].execute
+  before(:all) do
+    on hosts, '/opt/ood/nginx_stage/sbin/nginx_stage nginx_clean --force'
+    on hosts, 'mkdir -p /opt/hooks'
+    upload_portal_config('portal_with_prehook.yml')
+    update_ood_portal
+    restart_apache
+  end
+
+  after(:each) do
     browser.close
+    on hosts, '/opt/ood/nginx_stage/sbin/nginx_stage nginx_clean --force'
   end
 
-  it 'does not crash login when pre hook crashes' do
-    mnts = ["-v", "#{hook_fixture('bad_pre_hook')}:/opt/hooks/pun_pre_hook"]
-    mnts.concat ["-v", "#{portal_fixture('portal_with_prehook.yml')}:/etc/ood/config/ood_portal.yml"]
-    Rake::Task['test:start_test_container'].execute(mount_args: mnts)
+  context 'pre hook crash' do
+    before(:all) do
+      scp_to(hosts, hook_fixture('bad_pre_hook'), '/opt/hooks/pun_pre_hook')
+    end
 
-    browser_login(browser)
-    browser.goto ctr_base_url
-    expect(browser.title).to eq('Dashboard - Open OnDemand')
+    it 'does not crash login when pre hook crashes' do
+      browser_login(browser)
+      browser.goto ctr_base_url
+      expect(browser.title).to eq('Dashboard - Open OnDemand')
+    end
 
-    hook_out = container_exec("cat /tmp/hook.out")
-    expect(hook_out.chomp).to eq('/opt/hooks/pun_pre_hook: line 4: wont_find_cmd: command not found')
+    describe file('/tmp/hook.out') do
+      its(:content) { is_expected.to contain '/opt/hooks/pun_pre_hook: line 4: wont_find_cmd: command not found' }
+    end
   end
 
-  it 'does not crash login when pre hook is misconfigured' do
-    # no mount point for hook file
-    mnts = ["-v", "#{portal_fixture('portal_with_prehook.yml')}:/etc/ood/config/ood_portal.yml"]
-    Rake::Task['test:start_test_container'].execute(mount_args: mnts)
+  context 'missing pre hook' do
+    before(:all) do
+      on hosts, 'rm -f /opt/hooks/pun_pre_hook'
+    end
 
-    browser_login(browser)
-    browser.goto ctr_base_url
-    expect(browser.title).to eq('Dashboard - Open OnDemand')
+    it 'does not crash login when pre hook is misconfigured' do
+      browser_login(browser)
+      browser.goto ctr_base_url
+      expect(browser.title).to eq('Dashboard - Open OnDemand')
+    end
   end
 
-  it 'sets environment variables in the pre hook' do
-    mnts = ["-v", "#{hook_fixture('good_pre_hook')}:/opt/hooks/pun_pre_hook"]
-    mnts.concat ["-v", "#{portal_fixture('portal_with_prehook.yml')}:/etc/ood/config/ood_portal.yml"]
-    Rake::Task['test:start_test_container'].execute(mount_args: mnts)
+  context 'environment variables' do
+    before(:all) do
+      scp_to(hosts, hook_fixture('good_pre_hook'), '/opt/hooks/pun_pre_hook')
+    end
 
-    browser_login(browser)
-    browser.goto ctr_base_url
-    expect(browser.title).to eq('Dashboard - Open OnDemand')
-    hook_out = container_exec("cat /tmp/hook.out")
-    expect(hook_out).to eq("input args are --user ood\naccess token set\nemail claim set\n")
+    it 'does not crash login when pre hook crashes' do
+      browser_login(browser)
+      browser.goto ctr_base_url
+      expect(browser.title).to eq('Dashboard - Open OnDemand')
+    end
+
+    describe file('/tmp/hook.out') do
+      its(:content) { is_expected.to contain "input args are --user ood\naccess token set\nemail claim set\n" }
+    end
   end
 end
